Move performer interfaces to models
