+++ pkg/performer/query.go
@@ -7,15 +7,7 @@ import (
 	"github.com/stashapp/stash/pkg/models"
 )
 
-type Queryer interface {
-	Query(ctx context.Context, performerFilter *models.PerformerFilterType, findFilter *models.FindFilterType) ([]*models.Performer, int, error)
-}
-
-type CountQueryer interface {
-	QueryCount(ctx context.Context, galleryFilter *models.PerformerFilterType, findFilter *models.FindFilterType) (int, error)
-}
-
-func CountByStudioID(ctx context.Context, r CountQueryer, id int, depth *int) (int, error) {
+func CountByStudioID(ctx context.Context, r models.PerformerQueryer, id int, depth *int) (int, error) {
 	filter := &models.PerformerFilterType{
 		Studios: &models.HierarchicalMultiCriterionInput{
 			Value:    []string{strconv.Itoa(id)},
@@ -27,7 +19,7 @@ func CountByStudioID(ctx context.Context, r CountQueryer, id int, depth *int) (i
 	return r.QueryCount(ctx, filter, nil)
 }
 
-func CountByTagID(ctx context.Context, r CountQueryer, id int, depth *int) (int, error) {
+func CountByTagID(ctx context.Context, r models.PerformerQueryer, id int, depth *int) (int, error) {
 	filter := &models.PerformerFilterType{
 		Tags: &models.HierarchicalMultiCriterionInput{
 			Value:    []string{strconv.Itoa(id)},
@@ -39,7 +31,7 @@ func CountByTagID(ctx context.Context, r CountQueryer, id int, depth *int) (int,
 	return r.QueryCount(ctx, filter, nil)
 }
 
-func CountByAppearsWith(ctx context.Context, r CountQueryer, id int) (int, error) {
+func CountByAppearsWith(ctx context.Context, r models.PerformerQueryer, id int) (int, error) {
 	filter := &models.PerformerFilterType{
 		Performers: &models.MultiCriterionInput{
 			Value:    []string{strconv.Itoa(id)},
